requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.24.0

# Finnhub Client
finnhub-python==2.4.18

# Environment Variables
python-dotenv==1.0.0

# Validation
pydantic==2.6.0

# Reddit API
praw==7.7.1

# Optional but recommended for production
python-multipart==0.0.9